Move implementation to cpp files (#5151)
