+++ bcos-framework/bcos-framework/Common.h
@@ -18,6 +18,7 @@
  */
 #pragma once
 #include <bcos-utilities/Log.h>
+#include <csignal>
 
 #define METRIC LOG_BADGE("METRIC")
